+++ src/components/ToDoList/index.js
@@ -1,2 +1,3 @@
 export {ToDoList} from './ToDoList';
 export {ToDoForm} from './ToDoForm';
+export {ToDo} from './ToDo';
